requirements_bi_pe.txt
aiohttp
appdirs
APScheduler
astroid
async-timeout
attrs
Babel
backcall
bcrypt
botocore
cachetools
certifi
cffi
chardet
cliff
cmd2
colorama
cryptography
cycler
debtcollector
decorator
dnspython
dogpile.cache
eventlet
extras
fasteners
fixtures
freeze
fsspec
gcsfs
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-core
google-cloud-logging
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
greenlet
grpcio
httplib2
idna
importlib-metadata
ipython
ipython-genutils
iso8601
isort
jedi
jmespath
joblib
jsonpatch
jsonpointer
jsonschema
keystoneauth1
kiwisolver
lazy-object-proxy
linecache2
matplotlib
mccabe
monotonic
msgpack
multidict
munch
netaddr
netifaces
numpy
oauthlib
openstacksdk
os-brick
os-client-config
os-service-types
os-win
osc-lib
oslo.concurrency
oslo.config
oslo.context
oslo.i18n
oslo.log
oslo.privsep
oslo.serialization
oslo.service
oslo.utils
packaging
pandas
pandas-gbq
paramiko
parso
Paste
PasteDeploy
pbr
pexpect
pickleshare
Pillow
prettytable
prompt-toolkit
protobuf
psutil
ptyprocess
pyasn1
pyasn1-modules
pycparser
pydata-google-auth
Pygments
pylint
pymongo
PyMySQL
PyNaCl
pyOpenSSL
pyparsing
pyperclip
pyrsistent
python-cinderclient
python-dateutil
python-freezerclient
python-glanceclient
python-keystoneclient
python-mimeparse
python-neutronclient
python-novaclient
python-swiftclient
pytz
PyYAML
repoze.lru
requests
requests-oauthlib
requestsexceptions
rfc3986
Routes
rsa
scikit-learn
scipy
seaborn
simplejson
six
sklearn
stevedore
tenacity
testtools
threadpoolctl
toml
tqdm
traceback2
traitlets
typed-ast
typing-extensions
tzlocal
unittest2
uritemplate
urllib3
warlock
wcwidth
WebOb
wrapt
xlrd
yappi
yarl
zipp
pyarrow